modularized rest of game
