+++ index.html
@@ -2,5 +2,5 @@
 <script src="jspm_packages/system.js"></script>
 <script src="config.js"></script>
 <script>
-System.import('src/app.js');
+System.import('src/game.js');
 </script>
