+++ src/game.js
@@ -1,0 +1,3 @@
+import Startup from './startup';
+
+console.log(new Startup());
